requirements.txt
django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
django-celery==3.3.1
django-celery-beat==2.2.0
celery==4.4.2
requests==2.25.1
gunicorn==20.0.4

selenium==3.141.0
urllib3==1.25.8

soupsieve==2.1
beautifulsoup4==4.9.3
bs4==0.0.1

